+++ amethyst_animation/src/skinning/systems.rs
@@ -88,7 +88,8 @@ impl<'a> System<'a> for VertexSkinningSystem {
         {
             if let Some(global_inverse) = mesh_global.0.invert() {
                 let skin = skins.get(joint_transform.skin).unwrap();
-                let joint_matrices = skin.joints
+                let joint_matrices = skin
+                    .joints
                     .iter()
                     .map(|joint_entity| {
                         (